server/requirements.txt
# Core Flask dependencies
Flask==3.1.2
flask-cors==6.0.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
click==8.3.0
itsdangerous==2.2.0
blinker==1.9.0

# Google AI and Generative AI
google-generativeai==0.8.5
google-genai==1.42.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
proto-plus==1.26.1
protobuf==5.29.5

# Machine Learning and NLP
sentence-transformers==5.1.1
torch==2.8.0
transformers==4.57.0
huggingface-hub==0.35.3
tokenizers==0.22.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
numpy==2.3.3
joblib==1.5.2
threadpoolctl==3.6.0

# Vector Search and Similarity
faiss-cpu==1.9.0.post1

# LangChain for text processing
langchain==0.3.27
langchain-core==0.3.78
langchain-text-splitters==0.3.11
langsmith==0.4.33

# Text Processing and PDF
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.14.2
soupsieve==2.8
markdown==3.7.0

# OCR for scanned PDFs
pdf2image==1.17.0
pytesseract==0.3.13

# Note: pdf2image requires Poppler system dependency
# Install with: brew install poppler (macOS) or apt-get install poppler-utils (Ubuntu)

# HTTP and API clients
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10
h11==0.16.0
sniffio==1.3.1
anyio==4.11.0

# Data processing and validation
pydantic==2.12.0
pydantic_core==2.41.1
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# JSON and serialization
orjson==3.11.3
jsonpatch==1.33
jsonpointer==3.0.0
PyYAML==6.0.3

# Environment and configuration
python-dotenv==1.1.1
dotenv==0.9.9

# Database (if needed)
SQLAlchemy==2.0.43

# Utilities
tqdm==4.67.1
tenacity==9.1.2
packaging==25.0
setuptools==80.9.0
filelock==3.20.0
fsspec==2025.9.0
cachetools==6.2.0
regex==2025.9.18
zstandard==0.25.0

# WebSocket support
websockets==15.0.1

# Mathematical operations
sympy==1.14.0
mpmath==1.3.0
pyparsing==3.2.5

# Network and graph operations
networkx==3.5

# Image processing
pillow==11.3.0

# Authentication and security
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests-toolbelt==1.0.0
uritemplate==4.2.0

# Hugging Face specific
hf-xet==1.1.10
fpdf==1.7.2